Workflow module consolidation
