+++ backend/automationWorkflow/serializers.py
@@ -4,122 +4,114 @@ Handles data validation, transformation, and nested relationships.
 """
 from rest_framework import serializers
 from automationWorkflow.models import Workflow, WorkflowNode, WorkflowConnection
-from workflows.validators import ConnectionValidator
+from automationWorkflow.validators import ConnectionValidator
 from core.models import Project
 from django.core.exceptions import ValidationError as DjangoValidationError
 
 
 class WorkflowSerializer(serializers.ModelSerializer):
     """Serializer for Workflow model"""
-    
+
     project_id = serializers.IntegerField(required=False, allow_null=True)
     organization_id = serializers.IntegerField(read_only=True)
     created_by_id = serializers.IntegerField(read_only=True)
-    
+
     class Meta:
         model = Workflow
         fields = [
-            'id',
-            'name',
-            'description',
-            'project_id',
-            'organization_id',
-            'created_by_id',
-            'status',
-            'version',
-            'created_at',
-            'updated_at',
-            'is_deleted',
+            "id",
+            "name",
+            "description",
+            "project_id",
+            "organization_id",
+            "created_by_id",
+            "status",
+            "version",
+            "created_at",
+            "updated_at",
+            "is_deleted",
         ]
         read_only_fields = [
-            'id',
-            'organization_id',
-            'created_by_id',
-            'version',
-            'created_at',
-            'updated_at',
-            'is_deleted',
+            "id",
+            "organization_id",
+            "created_by_id",
+            "version",
+            "created_at",
+            "updated_at",
+            "is_deleted",
         ]
-    
+
     def validate_project_id(self, value):
         """
         Validate that project exists and belongs to the same organization
         as the requesting user (if provided).
         """
         if value is not None:
-            request = self.context.get('request')
+            request = self.context.get("request")
             qs = Project.objects.filter(id=value, is_deleted=False)
             if not qs.exists():
-                raise serializers.ValidationError(f'Project with id {value} does not exist')
+                raise serializers.ValidationError(f"Project with id {value} does not exist")
             if request and request.user and request.user.is_authenticated:
-                project = qs.select_related('organization').first()
+                project = qs.select_related("organization").first()
                 user_org_id = request.user.organization_id
                 if user_org_id and project.organization_id != user_org_id:
                     raise serializers.ValidationError(
-                        'Project must belong to the same organization as the creator.'
+                        "Project must belong to the same organization as the creator."
                     )
         return value
-    
+
     def create(self, validated_data):
         """
         Create workflow with organization / project / creator relationship.
-
-        - organization is derived from project.organization if project is set,
-          otherwise falls back to request.user.organization.
-        - created_by is always set to request.user when available.
         """
-        request = self.context.get('request')
-        project_id = validated_data.pop('project_id', None)
+        request = self.context.get("request")
+        project_id = validated_data.pop("project_id", None)
         project = None
 
         if project_id is not None:
             project = Project.objects.filter(id=project_id, is_deleted=False).select_related(
-                'organization'
+                "organization"
             ).first()
 
         organization = None
         if project is not None:
             organization = project.organization
         elif request and request.user and request.user.is_authenticated:
-            # Fallback to user's organization for org-global workflows
             organization = request.user.organization
 
-        # Attach foreign keys explicitly
         if project is not None:
-            validated_data['project'] = project
+            validated_data["project"] = project
         if organization is not None:
-            validated_data['organization'] = organization
+            validated_data["organization"] = organization
         if request and request.user and request.user.is_authenticated:
-            validated_data['created_by'] = request.user
+            validated_data["created_by"] = request.user
 
         return super().create(validated_data)
-    
+
     def update(self, instance, validated_data):
         """
         Update workflow with project relationship.
-
-        - project can be changed but must remain in the same organization.
-        - organization / created_by are not writable through this serializer.
         """
-        request = self.context.get('request')
-        project_id = validated_data.pop('project_id', None)
+        request = self.context.get("request")
+        project_id = validated_data.pop("project_id", None)
 
         if project_id is not None:
             project = Project.objects.filter(id=project_id, is_deleted=False).select_related(
-                'organization'
+                "organization"
             ).first()
             if not project:
                 raise serializers.ValidationError(
-                    {'project_id': f'Project with id {project_id} does not exist'}
+                    {"project_id": f"Project with id {project_id} does not exist"}
                 )
 
-            # Enforce same-organization constraint
             target_org_id = project.organization_id
             current_org_id = instance.organization_id
 
             if current_org_id and target_org_id != current_org_id:
                 raise serializers.ValidationError(
-                    {'project_id': 'Project must belong to the same organization as the workflow.'}
+                    {
+                        "project_id": "Project must belong to the same organization as the workflow."
+                    }
                 )
 
             if (
@@ -131,31 +123,36 @@ class WorkflowSerializer(serializers.ModelSerializer):
                 and target_org_id != request.user.organization_id
             ):
                 raise serializers.ValidationError(
-                    {'project_id': 'Project must belong to the same organization as the user.'}
+                    {"project_id": "Project must belong to the same organization as the user."}
                 )
 
-            validated_data['project'] = project
+            validated_data["project"] = project
 
-        # Prevent clients from changing organization / created_by explicitly
-        validated_data.pop('organization', None)
-        validated_data.pop('created_by', None)
+        validated_data.pop("organization", None)
+        validated_data.pop("created_by", None)
 
         return super().update(instance, validated_data)
 
 
 class WorkflowNodeSerializer(serializers.ModelSerializer):
     """Serializer for WorkflowNode model"""
-    
+
     workflow_id = serializers.IntegerField(read_only=True)
-    
+
     class Meta:
         model = WorkflowNode
         fields = [
-            'id', 'workflow_id', 'node_type', 'label', 'data',
-            'created_at', 'updated_at', 'is_deleted'
+            "id",
+            "workflow_id",
+            "node_type",
+            "label",
+            "data",
+            "created_at",
+            "updated_at",
+            "is_deleted",
         ]
-        read_only_fields = ['id', 'workflow_id', 'created_at', 'updated_at', 'is_deleted']
-    
+        read_only_fields = ["id", "workflow_id", "created_at", "updated_at", "is_deleted"]
+
     def validate_node_type(self, value):
         """Validate node type against allowed choices"""
         valid_types = [choice[0] for choice in WorkflowNode.NODE_TYPE_CHOICES]
@@ -164,16 +161,15 @@ class WorkflowNodeSerializer(serializers.ModelSerializer):
                 f'Invalid node type "{value}". Must be one of: {", ".join(valid_types)}'
             )
         return value
-    
+
     def validate_data(self, value):
         """Validate that data is a dictionary"""
         if not isinstance(value, dict):
-            raise serializers.ValidationError('Node data must be a dictionary/object')
+            raise serializers.ValidationError("Node data must be a dictionary/object")
         return value
-    
+
     def validate(self, attrs):
         """Additional validation for node creation"""
-        # Run model's clean method if updating existing instance
         if self.instance:
             node = self.instance
             for attr, value in attrs.items():
@@ -182,34 +178,49 @@ class WorkflowNodeSerializer(serializers.ModelSerializer):
                 node.clean()
             except DjangoValidationError as e:
                 raise serializers.ValidationError(e.message_dict)
-        
+
         return attrs
 
 
 class WorkflowNodeCreateSerializer(WorkflowNodeSerializer):
     """Serializer for creating workflow nodes"""
-    
+
     class Meta(WorkflowNodeSerializer.Meta):
-        fields = ['id', 'node_type', 'label', 'data', 'created_at', 'updated_at', 'is_deleted']
-        read_only_fields = ['id', 'created_at', 'updated_at', 'is_deleted']
+        fields = [
+            "id",
+            "node_type",
+            "label",
+            "data",
+            "created_at",
+            "updated_at",
+            "is_deleted",
+        ]
+        read_only_fields = ["id", "created_at", "updated_at", "is_deleted"]
 
 
 class WorkflowConnectionSerializer(serializers.ModelSerializer):
     """Serializer for WorkflowConnection model"""
-    
+
     workflow_id = serializers.IntegerField(read_only=True)
     source_node_id = serializers.IntegerField(required=False, allow_null=False)
     target_node_id = serializers.IntegerField(required=False, allow_null=False)
-    
+
     class Meta:
         model = WorkflowConnection
         fields = [
-            'id', 'workflow_id', 'source_node_id', 'target_node_id',
-            'connection_type', 'condition_config', 'priority',
-            'created_at', 'updated_at', 'is_deleted'
+            "id",
+            "workflow_id",
+            "source_node_id",
+            "target_node_id",
+            "connection_type",
+            "condition_config",
+            "priority",
+            "created_at",
+            "updated_at",
+            "is_deleted",
         ]
-        read_only_fields = ['id', 'workflow_id', 'created_at', 'updated_at', 'is_deleted']
-    
+        read_only_fields = ["id", "workflow_id", "created_at", "updated_at", "is_deleted"]
+
     def validate_connection_type(self, value):
         """Validate connection type against allowed choices"""
         valid_types = [choice[0] for choice in WorkflowConnection.CONNECTION_TYPE_CHOICES]
@@ -218,33 +229,30 @@ class WorkflowConnectionSerializer(serializers.ModelSerializer):
                 f'Invalid connection type "{value}". Must be one of: {", ".join(valid_types)}'
             )
         return value
-    
+
     def validate_condition_config(self, value):
         """Validate condition_config structure"""
         if not isinstance(value, dict):
-            raise serializers.ValidationError('Condition configuration must be a dictionary')
+            raise serializers.ValidationError("Condition configuration must be a dictionary")
         return value
-    
+
     def validate(self, attrs):
         """Validate connection rules"""
-        # Get workflow from context
-        workflow = self.context.get('workflow')
+        workflow = self.context.get("workflow")
         if not workflow:
-            raise serializers.ValidationError('Workflow context is required')
-        
-        # For updates, use existing values if not provided
+            raise serializers.ValidationError("Workflow context is required")
+
         if self.instance:
-            source_node_id = attrs.get('source_node_id', self.instance.source_node_id)
-            target_node_id = attrs.get('target_node_id', self.instance.target_node_id)
-            connection_type = attrs.get('connection_type', self.instance.connection_type)
-            condition_config = attrs.get('condition_config', self.instance.condition_config)
+            source_node_id = attrs.get("source_node_id", self.instance.source_node_id)
+            target_node_id = attrs.get("target_node_id", self.instance.target_node_id)
+            connection_type = attrs.get("connection_type", self.instance.connection_type)
+            condition_config = attrs.get("condition_config", self.instance.condition_config)
         else:
-            source_node_id = attrs.get('source_node_id')
-            target_node_id = attrs.get('target_node_id')
-            connection_type = attrs.get('connection_type', 'sequential')
-            condition_config = attrs.get('condition_config', {})
-        
-        # Validate that both nodes exist and belong to the workflow
+            source_node_id = attrs.get("source_node_id")
+            target_node_id = attrs.get("target_node_id")
+            connection_type = attrs.get("connection_type", "sequential")
+            condition_config = attrs.get("condition_config", {})
+
         try:
             source_node = WorkflowNode.objects.get(
                 id=source_node_id,
@@ -252,10 +260,12 @@ class WorkflowConnectionSerializer(serializers.ModelSerializer):
                 is_deleted=False,
             )
         except WorkflowNode.DoesNotExist:
-            raise serializers.ValidationError({
-                'source_node_id': f'Node with id {source_node_id} does not exist in this workflow'
-            })
-        
+            raise serializers.ValidationError(
+                {
+                    "source_node_id": f"Node with id {source_node_id} does not exist in this workflow"
+                }
+            )
+
         try:
             target_node = WorkflowNode.objects.get(
                 id=target_node_id,
@@ -263,59 +273,66 @@ class WorkflowConnectionSerializer(serializers.ModelSerializer):
                 is_deleted=False,
             )
         except WorkflowNode.DoesNotExist:
-            raise serializers.ValidationError({
-                'target_node_id': f'Node with id {target_node_id} does not exist in this workflow'
-            })
-        
-        # Validate connection rules using validator
+            raise serializers.ValidationError(
+                {
+                    "target_node_id": f"Node with id {target_node_id} does not exist in this workflow"
+                }
+            )
+
         try:
             ConnectionValidator.validate_connection_create(source_node, target_node, workflow)
         except DjangoValidationError as e:
             raise serializers.ValidationError(e.message_dict)
-        
-        # Validate conditional connection requirements
+
         if connection_type == WorkflowConnection.CONNECTION_TYPE_CONDITIONAL:
             if not condition_config:
-                raise serializers.ValidationError({
-                    'condition_config': 'Conditional connections require condition_config'
-                })
-        
-        # Validate loop connection requirements
+                raise serializers.ValidationError(
+                    {"condition_config": "Conditional connections require condition_config"}
+                )
+
         if connection_type == WorkflowConnection.CONNECTION_TYPE_LOOP:
             if not condition_config:
-                raise serializers.ValidationError({
-                    'condition_config': 'Loop connections require condition_config'
-                })
-            
-            max_iterations = condition_config.get('max_iterations')
+                raise serializers.ValidationError(
+                    {"condition_config": "Loop connections require condition_config"}
+                )
+
+            max_iterations = condition_config.get("max_iterations")
             if max_iterations is None:
-                raise serializers.ValidationError({
-                    'condition_config': 'Loop connections must specify max_iterations'
-                })
-            
+                raise serializers.ValidationError(
+                    {"condition_config": "Loop connections must specify max_iterations"}
+                )
+
             if not isinstance(max_iterations, int) or max_iterations < 1 or max_iterations > 1000:
-                raise serializers.ValidationError({
-                    'condition_config': 'max_iterations must be an integer between 1 and 1000'
-                })
-        
+                raise serializers.ValidationError(
+                    {
+                        "condition_config": "max_iterations must be an integer between 1 and 1000"
+                    }
+                )
+
         return attrs
 
 
 class WorkflowConnectionCreateSerializer(WorkflowConnectionSerializer):
     """Serializer for creating workflow connections"""
-    
+
     class Meta(WorkflowConnectionSerializer.Meta):
         fields = [
-            'id', 'source_node_id', 'target_node_id',
-            'connection_type', 'condition_config', 'priority',
-            'created_at', 'updated_at', 'is_deleted'
+            "id",
+            "source_node_id",
+            "target_node_id",
+            "connection_type",
+            "condition_config",
+            "priority",
+            "created_at",
+            "updated_at",
+            "is_deleted",
         ]
-        read_only_fields = ['id', 'created_at', 'updated_at', 'is_deleted']
+        read_only_fields = ["id", "created_at", "updated_at", "is_deleted"]
 
 
 class WorkflowGraphSerializer(serializers.Serializer):
     """Serializer for complete workflow graph (workflow + nodes + connections)"""
-    
+
     workflow = WorkflowSerializer(read_only=True)
     nodes = WorkflowNodeSerializer(many=True, read_only=True)
     connections = WorkflowConnectionSerializer(many=True, read_only=True)
@@ -323,31 +340,30 @@ class WorkflowGraphSerializer(serializers.Serializer):
 
 class BatchNodeOperationSerializer(serializers.Serializer):
     """Serializer for batch node operations"""
-    
+
     create = WorkflowNodeCreateSerializer(many=True, required=False)
     update = serializers.ListSerializer(child=serializers.DictField(), required=False)
     delete = serializers.ListSerializer(child=serializers.IntegerField(), required=False)
-    
+
     def validate_update(self, value):
         """Validate update operations have id field"""
         for item in value:
-            if 'id' not in item:
+            if "id" not in item:
                 raise serializers.ValidationError('Each update item must have an "id" field')
         return value
 
 
 class BatchConnectionOperationSerializer(serializers.Serializer):
     """Serializer for batch connection operations"""
-    
-    # Use DictField instead of nested serializer - validation happens in view with context
+
     create = serializers.ListSerializer(child=serializers.DictField(), required=False)
     update = serializers.ListSerializer(child=serializers.DictField(), required=False)
     delete = serializers.ListSerializer(child=serializers.IntegerField(), required=False)
-    
+
     def validate_update(self, value):
         """Validate update operations have id field"""
         for item in value:
-            if 'id' not in item:
+            if "id" not in item:
                 raise serializers.ValidationError('Each update item must have an "id" field')
         return value
